main: Make build_sbpf job succeed without programs specified
#### Problem

The build_sbpf job in the main workflow is a requirement to run tests,
but if a repo doesn't have any programs, the tests will never run. This
is an issue for the libraries repo, which has no programs.

#### Summary of changes

Move the `if` check from the top of the job to each step, so that
`build_sbpf` succeeds if no program packages are specified.